@ngnpope 2022-10-28
Closing in favour of #15405 which seems to be comparable or better, and includes some fixes.
Thanks for the extra test which I've taken the extra test over there.